pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kensa"
version = "0.1.0"
description = "The open source agent evals harness"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{name = "satyaborg", email = "satya.borg@gmail.com"}]
keywords = ["evals", "llm", "agents", "opentelemetry", "testing", "ai", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

[project.urls]
Homepage = "https://kensa.sh"
Repository = "https://github.com/satyaborg/kensa"
Issues = "https://github.com/satyaborg/kensa/issues"
Changelog = "https://github.com/satyaborg/kensa/releases"

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.40",
    "openinference-instrumentation-anthropic>=0.1",
]
openai = [
    "openai>=1.50",
    "openinference-instrumentation-openai>=0.1",
]
langchain = [
    "openinference-instrumentation-langchain>=0.1",
]
all = ["kensa[anthropic,openai,langchain]"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "ty>=0.0.24",
    "pytest-mock>=3.0",
    "pre-commit>=3.0",
    "langgraph>=0.6",
    "langchain-openai>=0.1",
]

[project.scripts]
kensa = "kensa.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/kensa"]

[tool.hatch.build.targets.sdist]
only-include = [
    "src/kensa",
    "pyproject.toml",
    "README.md",
    "CODE_OF_CONDUCT.md",
    "LICENSE",
    "CONTRIBUTING.md",
    "SECURITY.md",
]

[tool.ty.environment]
python-version = "3.10"
python = ".venv"

[tool.ty.src]
include = ["src", "tests"]

[tool.ty.analysis]
# Optional deps imported at runtime behind try/except
replace-imports-with-any = ["anthropic.**", "openai.**", "openinference.**"]

[tool.ty.rules]
all = "error"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires API keys (deselect with '-m \"not integration\"')",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
    "PT",   # flake8-pytest-style
    "PIE",  # flake8-pie (misc lints)
    "C4",   # flake8-comprehensions
    "RET",  # flake8-return
    "PERF", # perflint
]

[tool.ruff.lint.isort]
known-first-party = ["kensa"]
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"examples/**" = ["E501"]
